pyproject.toml
[project]
name = "youyou"
version = "0.1.0"
description = "本地智能助手系统 - 支持对话、记忆和任务管理"
readme = "README.md"
authors = [
    { name = "yaotutu", email = "957810668@qq.com" }
]
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.0",
    "langgraph>=1.0.0",
    "langchain-openai>=0.2.0",
    "langgraph-checkpoint-sqlite>=1.0.0",
    "mem0ai>=0.1.0",
    "python-dotenv>=1.0.0",
    "flask>=3.0.0",
    "flask-restx>=1.3.0",
    "flask-cors>=4.0.0",
]

[project.scripts]
youyou-server = "youyou.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"
